better formatting output pvpanel_print, fixed calculation of rs io and eg in pvpanel

+++ src/pvpanel/pvpanel.h
@@ -24,7 +24,7 @@ typedef struct pvPanel {
     double n;
     double rs;
     double rp;
-    double eg;
+    long double eg;
     long double io;
     double irr;
